pyproject.toml
[tool.poetry]
name = "basedosdados-api"
version = "0.0.1"
description = "Base dos Dados API"
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>", "Guilherme Peres <contact@gperes.dev>"]
license = "GPL-3.0-only"
homepage = "https://github.com/basedosdados/backend"
repository = "https://github.com/basedosdados/backend"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
django-health-check = "^3.17.0"
drf-yasg = "^1.21.4"
django-modeltranslation = "^0.18.7"
django-autoslug = "^1.9.8"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.18.1"
flake8 = "^4.0.1"

[tool.poetry.scripts]
lint = "scripts.lint:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
